project/requirements.txt
django==3.8
djangorestframework